requirements.txt
Jinja2
boto3
markdown
pyyaml